+++ templates/leptos-ssr/content/Cargo.toml.tmpl
@@ -17,7 +17,7 @@ leptos_wasi = { version = "0.1.3", optional = true }
 spin-sdk = { version = "3", optional = true }
 # `wasi` needs to be locked until https://github.com/fermyon/spin/issues/2928 lands in a release
 wasi = { version = "=0.13.2", optional = true }
-wasm-bindgen = { version = "0.2.96", optional = true }
+wasm-bindgen = { version = "0.2.99", optional = true }
 
 [workspace]
