+++ templates/leptos-ssr/content/src/app.rs
@@ -1,7 +1,11 @@
 use components::{Route, Router, Routes};
-use leptos::{prelude::*, task::spawn_local};
+use leptos::prelude::*;
 use leptos_meta::*;
 use leptos_router::*;
+
+use leptos::prelude::ElementChild;
+
+use crate::components::{home_page::HomePage, not_found::NotFound};
 
 #[cfg(feature = "ssr")]
 pub fn shell(options: LeptosOptions) -> impl IntoView {
@@ -11,11 +15,13 @@ pub fn shell(options: LeptosOptions) -> impl IntoView {
             <head>
                 <meta charset="utf-8"/>
                 <meta name="viewport" content="width=device-width, initial-scale=1"/>
+                <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet">
                 <AutoReload options=options.clone() />
-                <HydrationScripts options=options.clone() root="{{http-path}}"/>
+                <HydrationScripts options=options.clone() root=""/>
                 <MetaTags/>
             </head>
             <body>
+                <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"></script>
                 <App/>
             </body>
         </html>
@@ -30,7 +36,7 @@ pub fn App() -> impl IntoView {
     let fallback = || view! { "Page not found." }.into_view();
 
     view! {
-        <Stylesheet id="leptos" href="{{http-path}}/pkg/{{project-name | snake_case}}.css"/>
+        <Stylesheet id="leptos" href="/pkg/leptos_test.css"/>
         <Meta name="description" content="A website running its server-side as a WASI Component :D"/>
 
         <Title text="Welcome to Leptos X Spin!"/>
@@ -38,59 +44,10 @@ pub fn App() -> impl IntoView {
         <Router>
             <main>
                 <Routes fallback>
-                    <Route path=path!("{{http-path}}") view=HomePage/>
-                    <Route path=path!("{{http-path}}/*any") view=NotFound/>
+                    <Route path=path!("") view=HomePage />
+                    <Route path=path!("/*any") view=NotFound />
                 </Routes>
             </main>
         </Router>
     }
-}
-
-/// Renders the home page of your application.
-#[component]
-fn HomePage() -> impl IntoView {
-    // Creates a reactive value to update the button
-    let (count, set_count) = signal(0);
-    let on_click = move |_| {
-        set_count.update(|count| *count += 1);
-        spawn_local(async move {
-            save_count(count.get()).await.unwrap();
-        });
-    };
-
-    view! {
-      <h1>"Welcome to Leptos - served from Spin!"</h1>
-      <button on:click=on_click>"Click Me: " {count}</button>
-    }
-}
-
-/// 404 - Not Found
-#[component]
-fn NotFound() -> impl IntoView {
-    // set an HTTP status code 404
-    // this is feature gated because it can only be done during
-    // initial server-side rendering
-    // if you navigate to the 404 page subsequently, the status
-    // code will not be set because there is not a new HTTP request
-    // to the server
-    #[cfg(feature = "ssr")]
-    {
-        // this can be done inline because it's synchronous
-        // if it were async, we'd use a server function
-        if let Some(resp) = use_context::<leptos_wasi::response::ResponseOptions>() {
-            resp.set_status(leptos_wasi::prelude::StatusCode::NOT_FOUND);
-        }
-    }
-
-    view! { <h1>"Not Found"</h1> }
-}
-
-#[server(prefix = "{{http-path}}/api")]
-pub async fn save_count(count: u32) -> Result<(), ServerFnError<String>> {
-    println!("Saving value {count}");
-    let store = spin_sdk::key_value::Store::open_default().map_err(|e| e.to_string())?;
-    store
-        .set_json("{{project-name | snake_case}}_count", &count)
-        .map_err(|e| ServerFnError::ServerError(e.to_string()))?;
-    Ok(())
 }
