+++ templates/leptos-ssr/content/src/components/mod.rs
@@ -1,0 +1,5 @@
+use crate::prelude::*;
+
+pub mod home_page;
+pub mod not_found;
+
