+++ templates/leptos-ssr/content/src/components/not_found.rs
@@ -1,0 +1,23 @@
+use leptos::prelude::ElementChild;
+use leptos::*;
+use leptos::context::use_context;
+
+#[component]
+pub fn NotFound() -> impl IntoView {
+    // set an HTTP status code 404
+    // this is feature gated because it can only be done during
+    // initial server-side rendering
+    // if you navigate to the 404 page subsequently, the status
+    // code will not be set because there is not a new HTTP request
+    // to the server
+    #[cfg(feature = "ssr")]
+    {
+        // this can be done inline because it's synchronous
+        // if it were async, we'd use a server function
+        if let Some(resp) = use_context::<leptos_wasi::response::ResponseOptions>() {
+            resp.set_status(leptos_wasi::prelude::StatusCode::NOT_FOUND);
+        }
+    }
+
+    view! { <h1>"Not Found"</h1> }
+}
